Have all accounts be with total
Remove the distinction between if an account has total
value or not and instead going to change logic on the
backend where accounts with running totals (assets, liabilities)
show current running total as their total value.

That way all accounts have a total.

This removes the custom deserialization logic as well from dao